# Use object-hash to speed up cache key computation