--- a/package.json
+++ b/package.json
@@ -98,6 +98,7 @@
         "nodemon": "^2.0.20",
         "numeric": "^1.2.6",
         "oauth-signature": "^1.5.0",
+        "object-hash": "^3.0.0",
         "on-finished": "^2.4.1",
         "parse5": "^6.0.1",
         "passport": "^0.6.0",