--- a/package.json
+++ b/package.json
@@ -166,6 +166,7 @@
         "@types/mocha": "^10.0.1",
         "@types/node": "^18.11.18",
         "@types/node-fetch": "^2.6.2",
+        "@types/object-hash": "^3.0.2",
         "@types/passport": "^1.0.11",
         "@types/passport-azure-ad": "^4.3.1",
         "@types/pem": "^1.9.6",