--- a/question-servers/freeform.js
+++ b/question-servers/freeform.js
@@ -5,7 +5,6 @@ const fs = require('fs-extra');
 const path = require('path');
 const mustache = require('mustache');
 const cheerio = require('cheerio');
-const hash = require('crypto').createHash;
 const parse5 = require('parse5');
 const debug = require('debug')('prairielearn:' + path.basename(__filename, '.js'));
 const { instrumented } = require('@prairielearn/opentelemetry');