This should reduce synchronous thread-blocking work in prod. Note that the value of the cache key will change, so we'll temporarily get zero cache hits.

Benchmark:

```
cpu: Apple M1 Max
runtime: node v18.12.0 (arm64-darwin)

benchmark                                  time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------------------- -----------------------------
• big strings
----------------------------------------------------------------------------- -----------------------------
JSON.stringify + crypto.subtle.digest   68.53 ms/iter   (66.05 ms … 71.24 ms)  70.27 ms  71.24 ms  71.24 ms
JSON.stringify + crypto.createHash      61.59 ms/iter   (59.49 ms … 66.89 ms)  62.03 ms  66.89 ms  66.89 ms
object-hash                                11 ms/iter   (10.78 ms … 11.75 ms)  11.02 ms  11.75 ms  11.75 ms

summary for big strings
  object-hash
   5.6x faster than JSON.stringify + crypto.createHash
   6.23x faster than JSON.stringify + crypto.subtle.digest

• big object
----------------------------------------------------------------------------- -----------------------------
JSON.stringify + crypto.subtle.digest   63.03 ms/iter    (62.2 ms … 64.34 ms)  63.43 ms  64.34 ms  64.34 ms
JSON.stringify + crypto.createHash      58.03 ms/iter   (56.34 ms … 61.51 ms)  58.21 ms  61.51 ms  61.51 ms
object-hash                             10.08 ms/iter    (9.58 ms … 13.03 ms)  10.21 ms  13.03 ms  13.03 ms

summary for big object
  object-hash
   5.76x faster than JSON.stringify + crypto.createHash
   6.25x faster than JSON.stringify + crypto.subtle.digest
```

Benchmark source:

```ts
// @ts-check
const crypto = require('node:crypto');
const objectHash = require('object-hash');

const bigString = crypto.randomBytes(5 * 1024 * 1024).toString('hex');
const bigStrings = { foo: bigString, bar: bigString };

// Create a variable named `bigObject` that hsa 1000 keys, each of which is a
// string of 1000 bytes of random data.
const bigObject = {};
for (let i = 0; i < 1000; i++) {
  bigObject[i] = crypto.randomBytes(10000).toString('hex');
}

async function hashJsonDigest(obj) {
  const data = JSON.stringify(obj);
  const encoder = new TextEncoder();
  const digest = await crypto.subtle.digest('SHA-1', encoder.encode(data));
  Buffer.from(digest).toString('base64');
}

function hashJsonCreateHash(obj) {
  const data = JSON.stringify(obj);
  crypto.createHash('sha1').update(data).digest('base64');
}

function hashObjectHash(obj) {
  objectHash(obj, { algorithm: 'sha1', encoding: 'base64' });
}

async function main() {
  const { group, bench, run } = await import('mitata');

  group('big strings', () => {
    bench('JSON.stringify + crypto.subtle.digest', async () => hashJsonDigest(bigStrings));
    bench('JSON.stringify + crypto.createHash', () => hashJsonCreateHash(bigStrings));
    bench('object-hash', () => hashObjectHash(bigStrings));
  });

  group('big object', () => {
    bench('JSON.stringify + crypto.subtle.digest', async () => hashJsonDigest(bigObject));
    bench('JSON.stringify + crypto.createHash', () => hashJsonCreateHash(bigObject));
    bench('object-hash', () => hashObjectHash(bigObject));
  });

  await run();
}

main().catch((err) => {
  console.error(err);
  process.exit(1);
});
```